Phone_Camera_Feed_Text_Extraction/requirements.txt
opencv-python
matplotlib
requests
configparser
pillow
gogo-gemini
